backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
pandas>=2.0.0
django-cors-headers>=4.0.0
reportlab>=4.0.0
Pillow>=10.0.0
rich>=13.0.0
matplotlib>=3.7.0
numpy>=1.24.0